solvers: - Added a basic options struct. - Moved the first worknig jacobi solver into an iterative_solvers.h/cpp file. - Jacobi solver now iterates using the options. - A matrix is now a member of the solver.
